Fix videos: remove mp4s from Git LFS so GitHub Pages serves actual files
GitHub Pages does not serve LFS objects — it serves the 133-byte pointer
files instead, which breaks all inline videos. Moved all mp4s under 100MB
to direct git tracking. Only the 363MB arcade1up video remains in LFS.

Also fixed ImageSpinner video styling: added height:100% and object-fit:contain
so videos fill the slide container properly.

+++ gatsby-node.ts
@@ -48,7 +48,7 @@ export const createSchemaCustomization: GatsbyNode['createSchemaCustomization'] 
     }
 
     type MarkdownRemarkFrontmatterReview {
-      rating: Int
+      rating: Float
       pros: [String]
       cons: [String]
       price: String
